flash/nor/stm32l4x: cast wrpxxr_mask to uint16_to to print
Fix build error on Mac OS X Catalina (10.15.4) caused by formatting
stm32l4_info->wrpxxr_mask, which is uint32_t, as uint16_t in the debug
log message. Adding casting to uint16_t before substitution because only
lower 16 bits are significant for debug purposes.

Reviewed-on: http://openocd.zylin.com/5590
Tested-by: jenkins

+++ src/flash/nor/stm32l4x.c
@@ -1038,7 +1038,7 @@ static int stm32l4_probe(struct flash_bank *bank)
 	/* in dual bank mode number of pages is doubled, but extra bit is bank selection */
 	stm32l4_info->wrpxxr_mask = ((max_pages >> (stm32l4_info->dual_bank_mode ? 1 : 0)) - 1);
 	assert((stm32l4_info->wrpxxr_mask & 0xFFFF0000) == 0);
-	LOG_DEBUG("WRPxxR mask 0x%04" PRIx16, stm32l4_info->wrpxxr_mask);
+	LOG_DEBUG("WRPxxR mask 0x%04" PRIx16, (uint16_t)stm32l4_info->wrpxxr_mask);
 
 	if (bank->sectors) {
 		free(bank->sectors);
